start of setup for setup hooks
